refactor(docker-compose): split the wrapper into one module per domain
src/docker_compose.ts held the Compose-invocation detection, the settings
base, all fourteen command classes and the task object in one 944-line
file. Guideline 8 asks for one domain per file, and the file is about to
grow, so it splits first and gains commands second.

Each part moves verbatim into the module for its domain — invocation.ts
for detecting whether the host has the v2 plugin or the v1 standalone
binary, settings.ts for the shared base, lifecycle.ts for up, down,
start, stop, restart and rm, images.ts for build, pull and push, and
containers.ts for run, exec, logs, ps and config — leaving
docker_compose.ts as the assembler that composes DockerComposeTasks.

mod.ts now names its exports rather than re-exporting a single module
wholesale, so the package's public surface is stated in one place.

No argv, no validation and no signature changes travel with the move:
the existing suite passes as written, and the only edit to it is its
import path, which now points at the package entry point rather than at
one internal module.
